# openapi.yaml
openapi: 3.0.0
info:
  title: ShopBuilder API
  version: 2.0.0
  description: >-

    ## ShopBuilder — Enterprise Multi-Tenant E-Commerce Platform


    A scalable multi-tenant backend supporting multiple merchants, stores, and customers within isolated tenant
    namespaces.


    ### Role-Based Access Control (RBAC)


    | Role | Description |

    |------|-------------|

    | `SUPER_ADMIN` | Full platform access — manages tenants, users, everything |

    | `PLATFORM_ADMIN` | Manages merchants and platform configuration |

    | `MERCHANT_OWNER` | Manages own stores, products, and orders |

    | `STORE_MANAGER` | Manages inventory and orders for assigned store |

    | `CUSTOMER` | Places and views own orders |


    ### Authentication

    Include a Bearer token on all protected endpoints:

    ```

    Authorization: Bearer <access_token>

    ```


    Access tokens expire in **15 minutes**. Use `POST /auth/refresh` with your refresh token to get a new one.


    ### Multi-Tenancy

    All business data is scoped to a `tenantId`. It is embedded in the JWT and automatically applied.

    Admins can override `tenantId` via query parameter.
          
  contact:
    name: ShopBuilder API Support
    email: api@shopbuilder.io
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
...
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List products (tenant-scoped)
      parameters:
        - in: query
          name: storeId
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated product list
        '401':
          description: Unauthorized
    post:
      tags:
        - Products
      summary: Create a product (MERCHANT_OWNER, SUPER_ADMIN, PLATFORM_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Premium Cotton T-Shirt
                description:
                  type: string
                  example: 100% organic cotton, pre-shrunk
                price:
                  type: number
                  example: 29.99
                category:
                  type: string
                  example: Apparel
                storeId:
                  type: string
                  format: uuid
                tenantId:
                  type: string
                  format: uuid
                  description: Required for SUPER_ADMIN/PLATFORM_ADMIN
      responses:
        '201':
          description: Product created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID (with variants)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product with variants
        '404':
          description: Product not found
    patch:
      tags:
        - Products
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
      responses:
        '200':
          description: Updated product
        '404':
          description: Product not found
    delete:
      tags:
        - Products
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted
        '404':
          description: Product not found
  /products/{id}/variants:
    get:
      tags:
        - Products
      summary: List all variants for a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of variants
        '404':
          description: Product not found
  /products/{id}/variants/{variantId}:
    get:
      tags:
        - Products
      summary: Get a single variant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Variant detail
        '404':
          description: Variant not found
  /products/generate-variants:
    post:
      tags:
        - Products
      summary: Preview SKU matrix (cartesian product, not persisted)
      description: |
        Generates all attribute combinations. Example: sizes=["S","M"], colors=["Black","White"] → 4 SKUs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sizes:
                  type: array
                  items:
                    type: string
                  example:
                    - S
                    - M
                    - L
                colors:
                  type: array
                  items:
                    type: string
                  example:
                    - Black
                    - White
                    - Navy
                materials:
                  type: array
                  items:
                    type: string
                  example:
                    - Cotton
                    - Polyester
      responses:
        '200':
          description: Generated SKU combinations
          content:
            application/json:
              example:
                count: 18
                variants:
                  - sku: S-Black-Cotton
                    size: S
                    color: Black
                    material: Cotton
        '400':
          description: No attributes provided
  /products/save-variants:
    post:
      tags:
        - Products
      summary: Generate and persist variant SKU matrix for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
              properties:
                productId:
                  type: string
                  format: uuid
                sizes:
                  type: array
                  items:
                    type: string
                  example:
                    - S
                    - M
                    - L
                    - XL
                colors:
                  type: array
                  items:
                    type: string
                  example:
                    - Black
                    - White
                materials:
                  type: array
                  items:
                    type: string
                  example:
                    - Cotton
      responses:
        '201':
          description: Variants saved (idempotent — safe to re-run)
        '404':
          description: Product not found for this tenant
  /products/variants/{variantId}/stock:
    patch:
      tags:
        - Products
      summary: Adjust stock delta for a variant (legacy endpoint)
      description: Pass positive to add, negative to subtract. See /inventory for granular operations.
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  example: 50
      responses:
        '200':
          description: Updated variant
        '400':
          description: Insufficient stock
        '403':
          description: Forbidden
  /stores:
    get:
      tags:
        - Stores
      summary: List stores (tenant-scoped)
      parameters:
        - in: query
          name: merchantId
          schema:
            type: string
        - in: query
          name: tenantId
          schema:
            type: string
          description: Admin override
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated store list
    post:
      tags:
        - Stores
      summary: Create a store under a merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - merchantId
              properties:
                name:
                  type: string
                  example: Downtown Flagship
                description:
                  type: string
                  example: Main retail location
                merchantId:
                  type: string
                  format: uuid
                tenantId:
                  type: string
                  format: uuid
                  description: For SUPER_ADMIN/PLATFORM_ADMIN only
      responses:
        '201':
          description: Store created
        '404':
          description: Merchant not found
  /stores/{id}:
    get:
      tags:
        - Stores
      summary: Get store by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Store detail with product and order counts
        '404':
          description: Store not found
    patch:
      tags:
        - Stores
      summary: Update store name or description
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated store
    delete:
      tags:
        - Stores
      summary: Delete a store (SUPER_ADMIN, PLATFORM_ADMIN, MERCHANT_OWNER)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Store deleted
...